knowledge_distillation_project/hf_space_app/requirements.txt
gradio>=5.0.0
pandas>=2.0.0
torch>=2.1.0
transformers>=4.44.0
peft>=0.12.0
accelerate>=0.34.0
bitsandbytes>=0.43.0
evaluate>=0.4.3
rouge_score>=0.1.2
huggingface_hub>=0.25.0
sentencepiece>=0.2.0
llama-cpp-python>=0.3.1
